separates story into smaller file, adds helper dir
