+++ src/components/pager/pager.js
@@ -22,7 +22,7 @@ class Pager extends React.Component {
     /**
      * Total number of records
      */
-    totalRecords: PropTypes.string.isRequired,
+    totalRecords: PropTypes.number.isRequired,
 
     /**
      * Function called when any pager changes take place
